==> notebooks/0_Data_Preprocessing.ipynb <==
# %pip install pandas numpy matplotlib seaborn scikit-learn

# Dataset Building

import pandas as pd
import re

drug_df = pd.read_csv("../Datasets/Drug_related_tweets.csv")

drug_df = drug_df[["tweet"]]
drug_df.columns = ["text"]
drug_df["label"] = 1

print("Drug tweets:", len(drug_df))
display(drug_df.head(10))

medical_df = pd.read_csv("../Datasets/unique_medical_tweets_dataset.csv")

medical_df = medical_df[["text"]]

medical_df["label"] = 0

print("Medical tweets:", len(medical_df))
display(medical_df.head(10))


sentiment_df = pd.read_csv(
    "../Datasets/1.6-million-tweets-dataset.csv",
    encoding="latin-1",
    header=None
)

sentiment_df_0 = sentiment_df[sentiment_df[0] == 0]
sentiment_df_4 = sentiment_df[sentiment_df[0] == 4]
sentiment_df_0 = sentiment_df_0[[5]]
sentiment_df_0.columns = ["text"]
sentiment_df_4 = sentiment_df_4[[5]]
sentiment_df_4.columns = ["text"]

sentiment_df_0["label"] = 0
sentiment_df_4["label"] = 0

print("General tweets:", len(sentiment_df_0) , len(sentiment_df_4))
display(sentiment_df_0.head())
display(sentiment_df_4.head())


drug_df = drug_df.sample(20000, random_state=42)
medical_df = medical_df.sample(10000, random_state=42)
sentiment_df_0 = sentiment_df_0.sample(5000, random_state=42)
sentiment_df_4 = sentiment_df_4.sample(5000, random_state=42)

final_df = pd.concat([drug_df, medical_df, sentiment_df_0, sentiment_df_4], ignore_index=True)
final_df = final_df.sample(frac=1, random_state=42).reset_index(drop=True)

final_df.to_csv("../Datasets/final_dataset_raw.csv", index=False)

final_df["label"].value_counts()


# DATA CLEANING & PREPROCESSING

# %pip install nltk

import nltk
from nltk.corpus import stopwords
from nltk.stem import WordNetLemmatizer

nltk.download("stopwords")
nltk.download("wordnet")

stop_words = set(stopwords.words("english"))
lemmatizer = WordNetLemmatizer()


def clean_text(text):
    text = text.lower()
    text = re.sub(r"http\S+", "", text)
    text = re.sub(r"@\w+", "", text)
    text = re.sub(r"#\w+", "", text)
    text = re.sub(r"[^a-z\s]", "", text)
    text = text.strip()

    tokens = text.split()
    tokens = [lemmatizer.lemmatize(w) for w in tokens if w not in stop_words]

    return " ".join(tokens)


df = pd.read_csv("../Datasets/final_dataset_raw.csv")

df["clean_text"] = df["text"].apply(clean_text)

df = df[df["clean_text"].str.len() > 3]

df.drop(columns=["text"], inplace=True)

df.rename(columns={"clean_text": "text"}, inplace=True)

df.to_csv("../Datasets/final_dataset_cleaned.csv", index=False)

df.head()


df["label"].value_counts()
