==> notebooks/2_Model_Training_Baseline.ipynb <==
import numpy as np
import joblib
from scipy.sparse import load_npz

from sklearn.linear_model import LogisticRegression
from sklearn.svm import LinearSVC


X_train = load_npz("../Models/Features/X_train_tfidf.npz")
X_test  = load_npz("../Models/Features/X_test_tfidf.npz")

y_train = np.load("../Models/Features/y_train.npy")
y_test  = np.load("../Models/Features/y_test.npy")

print("Train shape:", X_train.shape)
print("Test shape :", X_test.shape)


# Logistic Regression

lr_model = LogisticRegression(
    max_iter=1000,
    class_weight="balanced",
    n_jobs=-1
)

lr_model.fit(X_train, y_train)


joblib.dump(
    lr_model,
    "../Models/logistic_regression_model.pkl"
)


# Linear SVM

svm_model = LinearSVC(
    class_weight="balanced"
)

svm_model.fit(X_train, y_train)


joblib.dump(
    svm_model,
    "../Models/linear_svm_model.pkl"
)
