==> notebooks/4_BERT_Data_Preparation.ipynb <==
# %pip install transformers torch datasets

import pandas as pd
import torch
from sklearn.model_selection import train_test_split
from transformers import BertTokenizer


df = pd.read_csv("../Datasets/final_dataset_cleaned.csv")

texts = df["text"].tolist()
labels = df["label"].tolist()

print(len(texts), len(labels))


X_temp, X_test, y_temp, y_test = train_test_split(
    texts,
    labels,
    test_size=0.15,
    random_state=42,
    stratify=labels
)

X_train, X_val, y_train, y_val = train_test_split(
    X_temp,
    y_temp,
    test_size=0.176,   # ≈ 15% of total
    random_state=42,
    stratify=y_temp
)

print(len(X_train), len(X_val), len(X_test))


import pandas as pd

train_df = pd.DataFrame({
    "text": X_train,
    "label": y_train
})

val_df = pd.DataFrame({
    "text": X_val,
    "label": y_val
})

test_df = pd.DataFrame({
    "text": X_test,
    "label": y_test
})


train_df.to_csv("../Models/Features/BERT/bert_train.csv", index=False)
val_df.to_csv("../Models/Features/BERT/bert_val.csv", index=False)
test_df.to_csv("../Models/Features/BERT/bert_test.csv", index=False)

print("BERT train/val/test files saved successfully.")
